BUG: Use double pow() to avoid platform-specific results.
The pow(1.0f, order) gave different results on Windows (where it
must have actually done a float power computation) versus Linux
where it was internally doing the computation with doubles.

+++ Charts/vtkAxis.cxx
@@ -596,9 +596,9 @@ double vtkAxis::CalculateNiceMinMax(double &min, double &max)
   double tickSpacing = range / maxTicks;
 
   int order = static_cast<int>(floor(log10(tickSpacing)));
-  double normTickSpacing = tickSpacing * pow(10.0f, -order);
+  double normTickSpacing = tickSpacing * pow(double(10.0), -order);
   double niceTickSpacing = this->NiceNumber(normTickSpacing, true);
-  niceTickSpacing *= pow(10.0f, order);
+  niceTickSpacing *= pow(double(10.0), order);
 
   if (isNegative)
     {
